task6: refactor - change names
